Server/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
english-words==1.0.3
gunicorn==20.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
